Plan and Build to get names into the FamilyMember data model.
The helper function should check online first, as the person could have updated their user name, and they may not exist in the local data. The local data should be the first fallback, if online or cache doesn't know, then final fallback is the Unknown.    What is a better way to do this lookup, caching Users seems like an extreme for something that won't be there often, but if that makes the most sense.

Error: Value of type 'Firestore' has no member 'getAllDocuments'

Not working, says Unknown User even after I just searched and added as a family member

Can we gather the names on entering the Family Screen, incase the Cache doesn't have the data yet?