Plan and build changes for minor issues
OK, I am able to leave a family but they don't get updated until they reopen the app.  Whats the best way to fix that.
I also see that Parent (Captain) sees invitations for all familes that exist.  So something is not being checked correctly.
I can also send multiple invites to the same user from the same Family.  It should be restricted on the search that if an invite is already sent that person does not show up.
I also do not see the invites on the invited persons UI.   Seems you are requring them to be in a family to see a invite to a family.  That's quite broken.

Build