Plan and build to see the invites....not sure they worked.
I don't see a way to see any pending family Invites.

Also instead of a prefix search can we search if the entire name/email (if public)/phone number (if public) CONTAINS the value?  It would be great if we had some detail text informing us how we found the user (userName, email, phone)

Lets add that we want a number bubble, telling us invites or pending friend requests/family memeber requets, etc. This is helpful for Family creators or people that use the sharecode so they can see that invite.